Autodelete old users as new are created
Also streamline all the old deletion queries to delete multiple items if
needed.

Oh, and Ctrl + Enter submits all forms.

Oh, and setup a migration system for that.

+++ models/article.js
@@ -82,11 +82,12 @@ module.exports = (sequelize) => {
           //    emptyTags.push(tag)
           //  }
           //}
-
-          article.sequelize.models.Tag.destroy({
-            where: { id: emptyTags.map(tag => tag.id) },
-            transaction: options.transaction,
-          })
+          if (emptyTags.length) {
+            article.sequelize.models.Tag.destroy({
+              where: { id: emptyTags.map(tag => tag.id) },
+              transaction: options.transaction,
+            })
+          }
         },
       },
       indexes: [
